requirements.txt
pandas
numpy
scikit-learn
matplotlib
torch
torchvision
torchaudio
tqdm
pyyaml
pytest
openpyxl
tabulate
